% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExtractAndWriteDissimilarityMatrix}
\alias{ExtractAndWriteDissimilarityMatrix}
\title{ExtractAndWriteDissimilarityMatrix}
\usage{
ExtractAndWriteDissimilarityMatrix(ifname, ofname, select, comment = "")
}
\arguments{
\item{ifname}{A string with the name of the file containing the original dissimilarity matrix in jmatrix format}

\item{ofname}{A string with the name of the file to contain the symmetric dissimilarity matrix between the selected rows.}

\item{select}{A boolean array with length equal to the size of the original matrix indicating (with true) which rows/columns must be extracted.}

\item{comment}{Comment to be atttached to the dissimilary matrix, added to the comment of the original one. Default: "" (no comment)}
}
\value{
No return value, called for side effects (creates a file)
}
\description{
Writes a binary symmetric matrix with some of the dissimilarities stored in a binary matrix in the scellpam package format.\cr
The purpose of this function is to get the dissimilarity matrix between some selected individuals of a larger set whose dissimilarity matrix has been
calculated before without calculating explicitly the dissimilarities again. The extracted matrix has the same data type as the original and of course
is a symmetric matrix, too. A comment can be added to the original matrix comment or set, if there was no previous comment.
}
\examples{
# TO BE DONE
}
