% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{JMatInfo}
\alias{JMatInfo}
\title{JMatInfo}
\usage{
JMatInfo(fname, fres = "")
}
\arguments{
\item{fname}{String with the file name that contains the binary data.}

\item{fres}{String with the name of the file to write the information. Default: "" (information is written to the console)}
}
\value{
No return value, called for its side effects (writes on screen or creates a file)
}
\description{
Shows in the screen or writes to a file information about a matrix stored in the binary format of package jmatrix
}
\examples{
Rf <- matrix(runif(48),nrow=6)
rownames(Rf) <- c("A","B","C","D","E","F")
colnames(Rf) <- c("a","b","c","d","e","f","g","h")
tmpfile1=paste0(tempdir(),"/Rfullfloat.bin")
JWriteBin(Rf,tmpfile1,dtype="float",dmtype="full",comment="Full matrix of floats")
JMatInfo(tmpfile1)
}
