% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{JMatInfoList}
\alias{JMatInfoList}
\title{JMatInfoList}
\usage{
JMatInfoList(fname)
}
\arguments{
\item{fname}{String with the file name that contains the binary data.}
}
\value{
A list with the following keys:
\itemize{
    \item mattype  - String with one of the values "full", "sparse", "symmetric" or "unknown"
    \item datatype - String with one of the values "uchar", "char", "ushort", "short", "uint", "int", "ulong", "long", "float", "double", "ldouble" or "unknown"
    \item endian   - String with one of the values "little","big"
    \item nrows    - Integer, the number of rows
    \item ncols    - Integer, the number of cols
    \item comment  - String with the comment, if stored, or the empty string otherwise
}
}
\description{
Returns a list with information about a matrix stored in the binary format of package jmatrix
}
\examples{
Rf <- matrix(runif(48),nrow=6)
rownames(Rf) <- c("A","B","C","D","E","F")
colnames(Rf) <- c("a","b","c","d","e","f","g","h")
tmpfile1=paste0(tempdir(),"/Rfullfloat.bin")
L <- JMatInfoList(tmpfile1)
cat(sep="","Matrix in file ",tmpfile1," is ",L$mattype,", has elements of type ")
cat(sep="",L$datatype," stored in ",L$endian," endian, its size is ",L$nrows," x ",L$ncols,"\n")
if (L$comment=="") { cat("Attached comment:\n",L$comment,"\n") }
}
