% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ScellpamGetDebug}
\alias{ScellpamGetDebug}
\title{ScellpamGetDebug}
\usage{
ScellpamGetDebug()
}
\value{
A list with the following keys:
\itemize{
     \item deb       - Boolean with the current debugging value of the scellpam (biological part) of this package
     \item debparpam - Boolean with the current debugging value of the parallel PAM part inside this package
     \item debjmat   - Boolean with the current debugging value of the jmatrix part inside this package
}
}
\description{
Obtains the current state of the debugging parameter. To be used by R functions which want to adjust their messages according to the general debugging settings.
}
\examples{
d<-ScellpamGetDebug()
}
