% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ScellpamSetDebug}
\alias{ScellpamSetDebug}
\title{ScellpamSetDebug}
\usage{
ScellpamSetDebug(deb = TRUE, debparpam = FALSE, debjmat = FALSE)
}
\arguments{
\item{deb}{boolean, TRUE to generate debug messages for the scellpam (biological part) of this package and FALSE to turn them off. Default: true}

\item{debparpam}{boolean, TRUE to generate debug messages for the parallel PAM part inside this package and FALSE to turn them off. Default: false}

\item{debjmat}{boolean, TRUE to generate debug messages for the jmatrix part inside this package and FALSE to turn them off. Default: false}
}
\value{
No return value, called for side effects (modification of internal variables)
}
\description{
Sets debugging in scellpam package to ON (with TRUE) or OFF (with FALSE) for several parts of it.\cr
On package load the default status is OFF.\cr
Setting debugging of any part to ON shows a message. Setting to OFF does not show anything (since debugging is OFF...)
}
\examples{
ScellpamSetDebug(TRUE,debparpam=FALSE,debjmat=FALSE)
ScellpamSetDebug(TRUE,debparpam=TRUE,debjmat=FALSE)
ScellpamSetDebug(TRUE,debparpam=TRUE,debjmat=TRUE)
}
